Add wallet creation txn to load test script
