+++ scripts/exclude_from_coverage.sh
@@ -16,5 +16,5 @@ EOF
 
 # Usage
 exclude_terms "mock" "c.out"
-exclude_terms "tss_payments_loadtest.go" "c.out"
+exclude_terms "tss_loadtest.go" "c.out"
 exclude_terms "fixtures.go" "c.out"
